Add files via upload
This project consists of Battery model which simulates the terminal voltage  and SOC of a high voltage battery pack. A Graphical User Interface asks for specifications of the battery pack and the parameters are fed into the simulink model which simulates the terminal voltage and SOC over a NEDC cycle. For computing the terminal voltage, two methods are used namely simple resistance and Thevenin first order model . The user will be asked for which method the model needs to use to compute the terminal voltage.